Problem: Send with don't wait or zero timeout failed with eagain when pipe is not full
Because don't wait or timeout zero rarely process commands, the send operation might fail even if the pipe is not full.
Recv method is already protecting again such a behavior. The following example will fail, although it is clear that the pipe is not full:

    void *ctx = zmq_ctx_new ();
    int timeo = 0;
    int hwm = 1;

    void *sb = zmq_socket (ctx, ZMQ_DEALER);
    zmq_setsockopt (sb, ZMQ_RCVHWM, &hwm, sizeof (int));
    zmq_bind (sb, "inproc://a");

    void *sc = zmq_socket (ctx, ZMQ_DEALER);
    zmq_setsockopt (sc, ZMQ_SNDHWM, &hwm, sizeof (int));
    zmq_setsockopt (sc, ZMQ_SNDTIMEO, &timeo, sizeof (int));
    zmq_connect (sc, "inproc://a");

    zmq_msg_t msg;
    zmq_msg_init (&msg);

    for (int i = 0; i < 10000; ++i) {
        int rc = zmq_msg_send (&msg, sc, 0);
        assert (rc >= 0);

        zmq_msg_recv (&msg, sb, 0);
    }

    zmq_msg_close (&msg);

    zmq_close (sb);
    zmq_close (sc);
    zmq_ctx_term (ctx);

Suggested solution: When send operation return EAGAIN with don't wait or zero timeout, process commands and try send again.